# swagger.yml
openapi: 3.0.3
info:
  title: Backend API example
  description: Backend API example API declaration
  termsOfService: http://swagger.io/terms/
  contact:
    email: wilmer.reluz.salvador@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
...
servers:
  - url: /api/v1
...
paths:
...
  /productos:
    get:
      description: Retorna una lista de todos los productos
      tags:
        - Productos
      operationId: GetAllProductos
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseAll"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Productos
      operationId: AddProducto
      summary: Crea un nuevo producto.
      description: Crea un nuevo producto y retorna la información del producto creado.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductoRequest"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductoResponse"
        204:
          description: No Content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /productos/{id}:
    get:
      tags:
        - Productos
      operationId: GetProductoById
      summary: Retorna información de un Producto.
      description: Retorna toda la informacion de un producto.
      parameters:
        - name: id
          in: path
          description: ID de Producto
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductoResponse"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags:
        - Productos
      operationId: UpdateProducto
      summary: Actualiza un producto.
      description: Actualiza un producto y retorna la información del producto actualizado.
      parameters:
        - name: id
          in: path
          description: ID de Producto
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductoRequest"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductoResponse"
        204:
          description: No Content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - Productos
      operationId: DeleteProducto
      summary: Elimina un nuevo producto de forma lógica.
      description: Elimina un producto y retorna la información del producto eliminado.
      parameters:
        - name: id
          in: path
          description: ID de Producto
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Updatet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductoResponse"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  #* Paths de Opción Producto
...
components:
  schemas:
...
    ResponseAll:
      type: object
      properties:
        success:
          type: boolean
        payload:
          type: array
        message:
          type: string
...
    Error:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        errors:
          type: array

    # ? Schema de Productos
    ProductoResponse:
      type: object
      properties:
        success:
          type: boolean
        payload:
          type: object
          properties:
            id:
              type: integer
            titulo:
              type: string
            descripcion:
              type: string
            url:
              type: string
            de_cocina:
              type: boolean
        message:
          type: string
    ProductoRequest:
      type: object
      properties:
        titulo:
          type: string
        descripcion:
          type: string
        url:
          type: string
        de_cocina:
          type: boolean
...